Trying to fix the API tests
