+++ src/api.routes.js
@@ -6,6 +6,7 @@ export const apiRoutes = [
   { method: "POST", path: "/api/users", config: userApi.create },
   { method: "DELETE", path: "/api/users", config: userApi.deleteAll },
   { method: "GET", path: "/api/users/{id}", config: userApi.findOne },
+  /*
   { method: "POST", path: "/api/categories", config: categoryApi.create },
   { method: "DELETE", path: "/api/categories", config: categoryApi.deleteAll },
   { method: "GET", path: "/api/categories", config: categoryApi.find },
@@ -13,5 +14,7 @@ export const apiRoutes = [
   { method: "DELETE", path: "/api/categories/{id}", config: categoryApi.deleteOne },
 
   { method: "POST", path: "/api/users/authenticate", config: userApi.authenticate },
+
+  */ 
 
 ];
