+++ src/api/user-api.js
@@ -1,32 +1,11 @@
 import Boom from "@hapi/boom";
 import { db } from "../models/db.js";
-import { UserSpec, UserSpecPlus, IdSpec, UserArray } from "../models/joi-schemas.js";
-import { validationError } from "./logger.js";
-import { createToken } from "./jwt-utils.js";
+
+// These are the Endpoints 
 
 export const userApi = {
-  authenticate: {
-    auth: false,
-    handler: async function (request, h) {
-      try {
-        const user = await db.userStore.getUserByEmail(request.payload.email);
-        if (!user) {
-          return Boom.unauthorized("User not found");
-        }
-        if (user.password !== request.payload.password) {
-          return Boom.unauthorized("Invalid password");
-        }
-        const token = createToken(user);
-        return h.response({ success: true, token: token }).code(201);
-      } catch (err) {
-        return Boom.serverUnavailable("Database Error");
-      }
-    },
-  },
   find: {
-    auth: {
-      strategy: "jwt",
-    },
+    auth: false,
     handler: async function (request, h) {
       try {
         const users = await db.userStore.getAllUsers();
@@ -35,16 +14,10 @@ export const userApi = {
         return Boom.serverUnavailable("Database Error");
       }
     },
-    tags: ["api"],
-    description: "Get all userApi",
-    notes: "Returns details of all userApi",
-    response: { schema: UserArray, failAction: validationError },
   },
 
   findOne: {
-    auth: {
-      strategy: "jwt",
-    },
+    auth: false,
     handler: async function (request, h) {
       try {
         const user = await db.userStore.getUserById(request.params.id);
@@ -56,11 +29,6 @@ export const userApi = {
         return Boom.serverUnavailable("No User with this id");
       }
     },
-    tags: ["api"],
-    description: "Get a specific user",
-    notes: "Returns user details",
-    validate: { params: { id: IdSpec }, failAction: validationError },
-    response: { schema: UserSpecPlus, failAction: validationError },
   },
 
   create: {
@@ -76,17 +44,10 @@ export const userApi = {
         return Boom.serverUnavailable("Database Error");
       }
     },
-    tags: ["api"],
-    description: "Create a User",
-    notes: "Returns the newly created user",
-    validate: { payload: UserSpec, failAction: validationError },
-    response: { schema: UserSpecPlus, failAction: validationError },
   },
 
   deleteAll: {
-    auth: {
-      strategy: "jwt",
-    },
+    auth: false,
     handler: async function (request, h) {
       try {
         await db.userStore.deleteAll();
@@ -95,8 +56,6 @@ export const userApi = {
         return Boom.serverUnavailable("Database Error");
       }
     },
-    tags: ["api"],
-    description: "Delete all userApi",
-    notes: "All userApi removed from Playtime",
   },
 };
+
