+++ test/api/user-api-test.js
@@ -3,18 +3,16 @@ import { poiService } from "./poi-service.js";
 import { assertSubset } from "../test-utils.js";
 import { maggie, testUsers } from "../fixtures.js";
 
-const users = new Array(testUsers.length);
-
-
 suite("User API tests", () => {
   setup(async () => {
-    await poiService.deleteAllUsers();
+      await poiService.deleteAllUsers();
     for (let i = 0; i < testUsers.length; i += 1) {
       // eslint-disable-next-line no-await-in-loop
-      users[0] = await poiService.createUser(testUsers[i]);
+      testUsers[i] = await poiService.createUser(testUsers[i]);
     }
   });
-  teardown(async () => {});
+  teardown(async () => {
+  });
 
   test("create a user", async () => {
     const newUser = await poiService.createUser(maggie);
@@ -22,7 +20,7 @@ suite("User API tests", () => {
     assert.isDefined(newUser._id);
   });
 
-  test("delete all userApi", async () => {
+  test("delete all users", async () => {
     let returnedUsers = await poiService.getAllUsers();
     assert.equal(returnedUsers.length, 3);
     await poiService.deleteAllUsers();
@@ -30,29 +28,8 @@ suite("User API tests", () => {
     assert.equal(returnedUsers.length, 0);
   });
 
-  test("get a user", async () => {
-    const returnedUser = await poiService.getUser(users[0]._id);
-    assert.deepEqual(users[0], returnedUser);
-  });
-
-  test("get a user - bad id", async () => {
-    try {
-      const returnedUser = await poiService.getUser("1234");
-      assert.fail("Should not return a response");
-    } catch (error) {
-      assert(error.response.data.message === "No User with this id");
-      // assert.equal(error.response.data.statusCode, 503);
-    }
-  });
-
-  test("get a user - deleted user", async () => {
-    await poiService.deleteAllUsers();
-    try {
-      const returnedUser = await poiService.getUser(users[0]._id);
-      assert.fail("Should not return a response");
-    } catch (error) {
-      assert(error.response.data.message === "No User with this id");
-      assert.equal(error.response.data.statusCode, 404);
-    }
+  test("get a user - success", async () => {
+    const returnedUser = await poiService.getUser(testUsers[0]._id);
+    assert.deepEqual(testUsers[0], returnedUser);
   });
 });
